requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
pydantic>=2.0
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sentry-sdk
pytest
httpx
email-validator
passlib[argon2]
